requirements.txt
Flask==2.3.2
langchain==0.0.260
openai==0.27.2
python-dotenv==1.0.0